Switching NewTable to Table (and OldTable) and fixing test issues
All instances of the old table are now called OldTable
in Cells and Components.

All instances of NewTable are now just called plain
Table in Cells and Components.

NOTE: We also fixed an issue whereby the chrome webdriver
for Selenium would not quit, resulting in possibly dozens
of orphanes processes when running test suites over and
over. This has now been resolved.
